Remove support for Go 1.8
